+++ issue_satisfaction_survey/models/__init__.py
@@ -2,5 +2,4 @@
 # © 2017 Mackilem Van der Laan, Trustcode
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-from . import models
-from . import wizard
+from . import project_issue
