--- issue_satisfaction_survey/security/ir.model.access.csv
@@ -1,2 +1,0 @@
-id,name,model_id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-access_wizard_all,access_wizard_all,project.issue.close,project_issue.grupo_cliente_externo,1,1,1,0
