+++ issue_satisfaction_survey/wizard/__init__.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
+# © 2017 Mackilem Van der Laan, Trustcode
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-import project_issue
+from . import project_issue_close
